+++ badge.js
@@ -1,5 +1,4 @@
-import { deleteLocalStorageData } from './deleteLocalStorageData.js';
-console.log('badge.js');
+const leetcode = 'https://leetcode.com';
 
 function updateBadgeText(state, tabId) {
   const text = state ? 'ON' : 'OFF';
@@ -10,36 +9,6 @@ chrome.runtime.onInstalled.addListener(() => {
   chrome.storage.local.set({ leetcodeAutoResetIsActive: true }, () => {
     chrome.action.setBadgeText({ text: 'ON' });
   });
-});
-
-const leetcode = 'https://leetcode.com';
-
-async function checkAndInjectContentScript(tabId, url) {
-  if (url.startsWith(leetcode)) {
-    const { leetcodeAutoResetIsActive } = await chrome.storage.local.get('leetcodeAutoResetIsActive');
-    if (!leetcodeAutoResetIsActive) return;
-
-    chrome.scripting.executeScript({
-      target: { tabId: tabId },
-      files: ['content.js']
-    });
-  }
-}
-
-chrome.tabs.onUpdated.addListener((tabId, changeInfo, tab) => {
-  if (changeInfo.status === 'complete' && tab.url) {
-    checkAndInjectContentScript(tabId, tab.url);
-  }
-});
-
-chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
-  if (request.action === 'deleteLocalStorage') {
-    console.log('deleteLocalStorage action received');
-    chrome.scripting.executeScript({
-      target: { tabId: sender.tab.id },
-      func: deleteLocalStorageData
-    });
-  }
 });
 
 async function toggleActivation(tab) {
