+++ content.js
@@ -1,4 +1,3 @@
-// content.js
 function checkForSolved() {
   const elements = document.querySelectorAll('.text-body.flex.flex-none.items-center.gap-1');
   for (let element of elements) {
