+++ service-worker.js
@@ -1,1 +1,2 @@
 import './badge.js';
+import './main.js';
